just admins can delete channels

+++ backend/api/src/chat/channel/channels.service.ts
@@ -1,15 +1,27 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import { ConflictException, Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
-import { Channel } from 'src/typeorm/entities/channel';
-import { CreateChannelDto } from 'src/dto/chat.dto';
+import { Channel } from 'src/typeorm';
+import { ChannelDto } from 'src/dto/chat.dto';
 import { UsersService } from 'src/users/users.service';
+import { CustomRequest } from 'src/utils/types';
+
 @Injectable()
 export class ChannelsService {
   constructor(
     @InjectRepository(Channel)
     private readonly channelsRepository: Repository<Channel>, private readonly usersService: UsersService
   ) {}
+
+  public async getChannels(userId: number): Promise<ChannelDto[]|null> {
+    let channels  = await this.channelsRepository
+    .createQueryBuilder().select()
+    .getMany()
+    const ret: ChannelDto[] = channels.map((channel) => {
+      return new ChannelDto(channel);
+    });
+    return ret;
+}
 
   findAll() {
     return this.channelsRepository.find({
@@ -28,24 +40,23 @@ export class ChannelsService {
   }
 
   
-  async create(createChannelDto: CreateChannelDto) {
-    //find and return error to front ?
+  async create(channelDto: ChannelDto) {
     const chan = await this.channelsRepository.findOne({
       where: {
-        name: createChannelDto.name
+        name: channelDto.name
       }
     });
-    if (!chan) {
-    createChannelDto.owner = await this.usersService.getOne(createChannelDto.owner.id);
-    const channel = this.channelsRepository.create(createChannelDto);
+    if (!chan && channelDto.name.length > 2) {
+    channelDto.owner = await this.usersService.getOne(channelDto.owner.id);
+    const channel = this.channelsRepository.create(channelDto);
     return this.channelsRepository.save(channel);
     }
     else
-      console.error("channel name already taken");
+      console.error("channel name invalide or already taken");
       
   }
 
-  async update(id: string, updateChannelDto: CreateChannelDto) { 
+  async update(id: string, updateChannelDto: ChannelDto) { 
     const channel = await this.channelsRepository.preload({
       id: +id,
       ...updateChannelDto,
@@ -56,10 +67,13 @@ export class ChannelsService {
     return this.channelsRepository.save(channel);
   }
 
-  async remove(id: string) { 
+  async remove(userID: number ,id: string) {
     const channel = await this.findOne(id);
     if (!channel) {
       throw new NotFoundException(`Channel [${id}] not found`);
+    }
+    if (channel.admin.some((admin) => {return admin.id == userID})) {
+      throw new NotFoundException(`User not found in the admin data`);
     }
     return this.channelsRepository.remove(channel);
   }
